+++ src/main/java/app/AppBuilder.java
@@ -3,6 +3,12 @@ package app;
 import data_access.FileUserDataAccessObject;
 import entity.UserFactory;
 import interface_adapter.ViewManagerModel;
+import interface_adapter.choose_fav_pokemon.ChooseFavPokemonController;
+import interface_adapter.choose_fav_pokemon.ChooseFavPokemonPresenter;
+import interface_adapter.choose_fav_pokemon.ChooseFavPokemonViewModel;
+import interface_adapter.edit_profile.EditProfileController;
+import interface_adapter.edit_profile.EditProfilePresenter;
+import interface_adapter.edit_profile.EditProfileViewModel;
 import interface_adapter.logged_in.*;
 import interface_adapter.login.LoginController;
 import interface_adapter.login.LoginPresenter;
@@ -14,6 +20,9 @@ import interface_adapter.signup.SignupViewModel;
 import use_case.change_password.ChangePasswordInputBoundary;
 import use_case.change_password.ChangePasswordInteractor;
 import use_case.change_password.ChangePasswordOutputBoundary;
+import use_case.choose_fav_pokemon.ChooseFavPokemonInputBoundary;
+import use_case.choose_fav_pokemon.ChooseFavPokemonInteractor;
+import use_case.choose_fav_pokemon.ChooseFavPokemonOutputBoundary;
 import use_case.customization.settings.*;
 import use_case.edit_profile.*;
 import use_case.login.LoginInputBoundary;
