+++ src/main/java/interface_adapter/choose_fav_pokemon/ChooseFavPokemonViewModel.java
@@ -1,4 +1,4 @@
-package interface_adapter.logged_in;
+package interface_adapter.choose_fav_pokemon;
 
 import entity.User;
 import interface_adapter.ViewModel;
