+++ src/main/java/interface_adapter/edit_profile/EditProfileController.java
@@ -1,4 +1,4 @@
-package interface_adapter.logged_in;
+package interface_adapter.edit_profile;
 
 import use_case.edit_profile.EditProfileInputBoundary;
 import use_case.edit_profile.EditProfileInputData;
