+++ src/main/java/interface_adapter/edit_profile/EditProfileState.java
@@ -1,4 +1,4 @@
-package interface_adapter.logged_in;
+package interface_adapter.edit_profile;
 
 import entity.User;
