+++ src/main/java/use_case/choose_fav_pokemon/ChooseFavPokemonInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.edit_profile;
+package use_case.choose_fav_pokemon;
 
 /**
  * The Choose Favourite Pokemon Use Case.
